requirements.txt
absl-py==0.9.0
appdirs==1.4.3
asgiref==3.2.7
astor==0.8.1
boto3==1.12.36
botocore==1.15.36
bson==0.5.8
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coverage==5.0.4
cryptography==2.8
cycler==0.10.0
dataclasses==0.7
DateTime==4.3
decorator==4.4.2
distlib==0.3.0
dj-database-url==0.4.2
Django==2.2.12
django-cors-headers==3.2.1
django-cors-middleware==1.3.1
django-environ==0.4.5
django-extensions==2.2.9
django-filter==1.0.4
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djongo==1.3.2
dnspython==1.16.0
docutils==0.15.2
drf-extensions==0.6.0
drf-nested-routers==0.90.0
filelock==3.0.12
FormEncode==1.3.1
gast==0.3.3
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gspread==3.6.0
gunicorn==19.7.1
h5py==2.10.0
httpie==0.9.9
httplib2==0.17.3
idna==2.9
image==1.5.28
imageio==2.8.0
imageio-ffmpeg==0.4.1
importlib-metadata==1.6.0
importlib-resources==1.4.0
imutils==0.5.3
jmespath==0.9.5
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.0
pandas==1.0.3
Pillow==7.0.0
pipenv==2018.11.26
protobuf==3.11.3
psutil==5.7.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.2.0
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.6
pyperclip==1.8.0
pytesseract==0.3.3
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
PyWavelets==1.1.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.2.4
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post2
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.15
virtualenv-clone==0.5.4
Werkzeug==1.0.1
whitenoise==3.3.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
